Time: 0 ms (100%), Space: 8.6 MB (89.66%) - LeetHub

+++ 3536-maximum-product-of-two-digits/3536-maximum-product-of-two-digits.cpp
@@ -16,6 +16,6 @@ public:
             }
         }
 
-        return int(fh*sh);
+        return (fh*sh);
     }
 };
